Removed all references to "togglearmtest"
