+++ src/org/usfirst/frc/team166/robot/OI.java
@@ -3,8 +3,6 @@ package org.usfirst.frc.team166.robot;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.JoystickButton;
 
-import org.usfirst.frc.team166.robot.commands.LoadingProcess;
-import org.usfirst.frc.team166.robot.commands.togglearmtest;
 import org.usfirst.frc.team166.robot.commands.drive.DriveWithGyro;
 import org.usfirst.frc.team166.robot.commands.drive.HighGear;
 import org.usfirst.frc.team166.robot.commands.drive.LowGear;
@@ -12,6 +10,7 @@ import org.usfirst.frc.team166.robot.commands.drive.Neutral;
 import org.usfirst.frc.team166.robot.commands.intake.IntakeMotorForward;
 import org.usfirst.frc.team166.robot.commands.intake.IntakeMotorReverse;
 import org.usfirst.frc.team166.robot.commands.intake.IntakeMotorStop;
+import org.usfirst.frc.team166.robot.commands.intake.LoadingProcess;
 import org.usfirst.frc.team166.robot.commands.intake.ToggleIntakeSolenoid;
 import org.usfirst.frc.team166.robot.commands.shooter.Aim;
 
@@ -61,7 +60,6 @@ public class OI {
 		CPbutton3.whenReleased(new IntakeMotorStop());
 		CPbutton4.whenPressed(new ToggleIntakeSolenoid());
 		CPbutton5.whenPressed(new LoadingProcess());// this the entire loading and prepping process.
-		CPbutton6.whenPressed(new togglearmtest());
 
 		// The Following commands are mapped from buttons on a joystick and may need to be changed if the
 		// copiolits controller turns out to be an Xbox controler
